+++ shared/src/index.ts
@@ -11,6 +11,8 @@ export * from "./gkPairing";
 export * from "./probableLineup";
 export * from "./setPieceTaker";
 export * from "./manager";
+export * from "./teamPref";
+export * from "./teamPreferences";
 export * from "./user";
 export * from "./avatar";
 export * from "./purchase";